# target/classes/assets/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "File Reader Rest API"
  },
  "host" : "localhost:4444",
  "tags" : [ {
    "name" : "phase 1 - category"
  }, {
    "name" : "phase 1 - data"
  }, {
    "name" : "phase 1 - document"
  }, {
    "name" : "phase 1 - search"
  }, {
    "name" : "phase 1 - system"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/api/categories" : {
      "get" : {
        "tags" : [ "phase 1 - category" ],
        "summary" : "API-01 | Get list Category",
        "description" : "",
        "operationId" : "getListCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "country",
          "in" : "query",
          "description" : "The country value",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mobileid",
          "in" : "query",
          "description" : "The mobileid value",
          "required" : true,
          "type" : "string",
          "default" : "167e5a47e70ee6c3_sdk29_tg28"
        }, {
          "name" : "token",
          "in" : "query",
          "description" : "The token value",
          "required" : true,
          "type" : "string",
          "default" : "testtoken"
        }, {
          "name" : "appid",
          "in" : "query",
          "description" : "The appid value",
          "required" : true,
          "type" : "string",
          "default" : "testappid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Return Category list is OK.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseBase"
              }
            }
          },
          "400" : {
            "description" : "Failed to return the Category list."
          }
        }
      }
    },
    "/api/data" : {
      "get" : {
        "tags" : [ "phase 1 - data" ],
        "summary" : "API-0101 | Return list of popular artwork",
        "description" : "pageNumber default value is 1 ",
        "operationId" : "getPopularArtwork",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Return Data list is OK.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Data"
              }
            }
          },
          "400" : {
            "description" : "Failed to return the data list."
          }
        }
      }
    },
    "/api/documents" : {
      "get" : {
        "tags" : [ "phase 1 - document" ],
        "summary" : "API-02 | Return list document of category",
        "description" : "",
        "operationId" : "getListDocumentOfCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "cateid",
          "in" : "query",
          "description" : "The id category value",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "country",
          "in" : "query",
          "description" : "The country value",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The limit value",
          "required" : true,
          "type" : "integer",
          "default" : 100,
          "minimum" : 0.0,
          "format" : "int32"
        }, {
          "name" : "mobileid",
          "in" : "query",
          "description" : "The mobileid value",
          "required" : true,
          "type" : "string",
          "default" : "167e5a47e70ee6c3_sdk29_tg28"
        }, {
          "name" : "token",
          "in" : "query",
          "description" : "The token value",
          "required" : true,
          "type" : "string",
          "default" : "testtoken"
        }, {
          "name" : "appid",
          "in" : "query",
          "description" : "The appid value",
          "required" : true,
          "type" : "string",
          "default" : "testappid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Return Data list is OK.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseBase"
              }
            }
          },
          "404" : {
            "description" : "Failed to return the data list."
          }
        }
      }
    },
    "/api/documents/record-action" : {
      "post" : {
        "tags" : [ "phase 1 - document" ],
        "summary" : "API-06 | Record Action File",
        "description" : "",
        "operationId" : "recordActionFile",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "header",
          "description" : "The document value",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "lang",
          "in" : "header",
          "description" : "The lang value",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "header",
          "description" : "The type value",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mobileid",
          "in" : "header",
          "description" : "The mobileid value",
          "required" : true,
          "type" : "string",
          "default" : "167e5a47e70ee6c3_sdk29_tg28"
        }, {
          "name" : "token",
          "in" : "header",
          "description" : "The token value",
          "required" : true,
          "type" : "string",
          "default" : "testtoken"
        }, {
          "name" : "appid",
          "in" : "header",
          "description" : "The appid value",
          "required" : true,
          "type" : "string",
          "default" : "testappid"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Status"
              }
            }
          },
          "400" : {
            "description" : "unsuccess"
          }
        }
      }
    },
    "/api/reindex/keyword" : {
      "get" : {
        "tags" : [ "phase 1 - search" ],
        "summary" : "API-10-05 | Rebuild indexing for keyword or search hashtag",
        "description" : "",
        "operationId" : "rebuildKeyword",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Number of deleted and inserted document is OK.",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          },
          "400" : {
            "description" : "Failed to delete and insert document is OK."
          }
        }
      }
    },
    "/api/search" : {
      "get" : {
        "tags" : [ "phase 1 - search" ],
        "summary" : "API-10-03 | Fulltext search keyword ",
        "description" : "",
        "operationId" : "searchKeyword",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "lang",
          "in" : "query",
          "description" : "The country value",
          "required" : true,
          "type" : "string",
          "default" : "US"
        }, {
          "name" : "q",
          "in" : "query",
          "description" : "The keyword value",
          "required" : true,
          "type" : "string",
          "default" : "hello"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "The offset value",
          "required" : true,
          "type" : "integer",
          "default" : 0,
          "minimum" : 0.0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The limit value",
          "required" : true,
          "type" : "integer",
          "default" : 100,
          "minimum" : 0.0,
          "format" : "int32"
        }, {
          "name" : "mobileid",
          "in" : "query",
          "description" : "The mobileid value",
          "required" : true,
          "type" : "string",
          "default" : "167e5a47e70ee6c3_sdk29_tg28"
        }, {
          "name" : "token",
          "in" : "query",
          "description" : "The token value",
          "required" : true,
          "type" : "string",
          "default" : "testtoken"
        }, {
          "name" : "appid",
          "in" : "query",
          "description" : "The appid value",
          "required" : true,
          "type" : "string",
          "default" : "testappid"
        }, {
          "name" : "searchtype",
          "in" : "query",
          "description" : "The search type value",
          "required" : true,
          "type" : "string",
          "default" : "keyword"
        } ],
        "responses" : {
          "200" : {
            "description" : "Return Wallpaper list is OK.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Response"
              }
            }
          },
          "400" : {
            "description" : "Failed to return the Wallpaper list."
          }
        }
      }
    },
    "/api/system/heartbeat" : {
      "head" : {
        "tags" : [ "phase 1 - system" ],
        "summary" : "Allows the HAProxy to check the server OK or NOK.",
        "description" : "",
        "operationId" : "ping",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/api/system/logLevel" : {
      "get" : {
        "tags" : [ "phase 1 - system" ],
        "summary" : "Log level 1: INFO, 2: WARNING 3: DEBUG",
        "description" : "",
        "operationId" : "log",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "level",
          "in" : "query",
          "description" : "level",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Data" : {
      "type" : "object",
      "properties" : {
        "createdDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        },
        "idJob" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Link" : {
      "type" : "object"
    },
    "Pagination" : {
      "type" : "object",
      "properties" : {
        "pageNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pageSize" : {
          "type" : "integer",
          "format" : "int32"
        },
        "total" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Response" : {
      "type" : "object",
      "properties" : {
        "status" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "hasnext" : {
          "type" : "boolean",
          "default" : false
        },
        "nextoffset" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "object"
        },
        "pagination" : {
          "$ref" : "#/definitions/Pagination"
        },
        "links" : {
          "$ref" : "#/definitions/Link"
        }
      }
    },
    "ResponseBase" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string"
        },
        "status" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "object"
        },
        "pagination" : {
          "$ref" : "#/definitions/Pagination"
        },
        "links" : {
          "$ref" : "#/definitions/Link"
        }
      }
    },
    "Status" : {
      "type" : "object",
      "properties" : {
        "statusCode" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}